requirements.txt
discord.py==2.4.0
fastapi==0.103.1
uvicorn==0.23.0